# GODRIVER-3132 Use dst buf pool for zstd encoding.