@qingyang-hu 2024-06-26
> Does this need to be ported to `master`?

Yes